backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.135.1",
    "numpy>=2.4.2",
    "pillow>=12.1.1",
    "pydicom>=3.0.1",
    "python-multipart>=0.0.22",
    "uvicorn>=0.41.0",
    "zstandard>=0.25.0",
]
